html/semantics/embedded-content/media-elements/playing-the-media-resource/pause-move-to-other-document.html

<!doctype html>
<title>paused state when moving to other document</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<div id="log"></div>
<video hidden></video>
<iframe hidden></iframe>
<script>
async_test(function(t) {
  var v = document.querySelector('video');
  v.src = getVideoURI('/media/movie_300');
  v.play();
  v.onplaying = t.step_func(function() {
    assert_false(v.paused, 'paused after playing');
    document.querySelector('iframe').contentDocument.body.appendChild(v);

    // Video should be paused because player reuse is not currently enabled
    // for non-pip related move.
    assert_true(v.paused, 'paused after moving');
    t.step_timeout(function() {
      assert_true(v.paused, 'paused after stable state')
      t.done();
    }, 0);
  });
});
</script>
